+++ bin/server.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
 
-require('./collaborate-editing.server')
-require('./note-synchronizer')
+require('./collaborate-editing-server')
+require('./note-synchronizer-server')
